fix: Added drag constraint to repeat until modal [SAMPLER-94]
This moves the drag constraint code from the variable setting modal to a hook which is then used in both the variable setting and repeat until modals.

With this change touch support is also added to the repeat until modal drag handling.
